requirements.txt
ccxt>=4.0.0
python-telegram-bot>=20.0
websockets>=12.0
aiohttp>=3.9
apscheduler>=3.10
python-dotenv>=1.0
loguru>=0.7
